backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
prophet==1.1.5
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.4
supabase==2.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
google-generativeai==0.4.1
requests==2.31.0
scipy==1.12.0
